% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectData.R
\name{addZeroes}
\alias{addZeroes}
\title{Add zeroes for non-observed species in ecological data}
\usage{
addZeroes(data, level = "all")
}
\arguments{
\item{data}{A dataframe with 'stationname', 'scientificname' and 'observationdate' columns. For example, the result
of \code{\link[=importBioticData]{importBioticData()}}.}

\item{level}{Character. Either \code{all} or \code{station}.}
}
\value{
A dataframe
}
\description{
Add zeroes for non-observed species in ecological data. \code{addZeroes()} sets a unique list of scientific names and add
zeroes for each of the species not found in a station. In this way, the absence of a specie in a station can be taken
into account.
}
\examples{
# make a list of the dataset ids
imis_datasets <- c(1073, 5191)

# download data
biotic_data_imis <- importBioticData(imis_datasets, 2015, 2021, 'imis')

# add zeroes
biotic_data_zeroes <- addZeroes(biotic_data_imis, level = "station")
}
