% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporalAggregating.R
\name{aggregateMonthlyPerStation}
\alias{aggregateMonthlyPerStation}
\title{Aggregate data monthly per station}
\usage{
aggregateMonthlyPerStation(data, stat = "mean")
}
\arguments{
\item{data}{A dataframe with \code{datetime}, \code{stationname} and \code{value} columns. For example, the result.
of \code{\link[=importAbioticData]{importAbioticData()}} or \code{\link[=importBioticData]{importBioticData()}}}

\item{stat}{The statistics to aggregate data. Predifined option is "\code{mean}". See Details for the full list.}
}
\value{
A dataframe.
}
\description{
Aggregate data monthly per station.
}
\details{
Available statistics are:
\itemize{
\item "\code{mean}" calculates Arithmetic Mean, uses \code{\link[base:mean]{base::mean()}},
\item "\code{sd}" calculates Stadard Deviation, uses \code{\link[stats:sd]{stats::sd()}}
\item "\code{se}" calculates Standard Error,
\item "\code{median}" calculates the Median Value. uses \code{\link[stats:median]{stats::median()}},
\item "\code{min}" calculates the minima, uses \code{\link[base:Extremes]{base::min()}},
\item "\code{max}" calculates the maxima, uses \code{\link[base:Extremes]{base::max()}},
\item "\verb{5perc}" calculates the 5th percentile,
\item "\verb{10perc}" calculates the 10th percentile,
\item "\verb{90perc}" calculates the 90th percentile,
\item "\verb{99perc}" calculates the 99th percentile,
\item "\code{sum}" calculates the sum, uses \code{\link[base:sum]{base::sum()}}
}
}
\examples{
# import data from high- and lowtide parameters
tide_parameters <- c(9694,9695)

# download data
abiotic_data <- importAbioticData(tide_parameters, start = 2012, end = 2015)

# calculate monthly mean per station
abiotic_data_median <- aggregateMonthlyPerStation(abiotic_data, stat = "median")
}
