% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporalAggregating.R
\name{aggregatePeriodPerZone}
\alias{aggregatePeriodPerZone}
\title{Aggregate data for year range per zone (level 3)}
\usage{
aggregatePeriodPerZone(data, stat = "mean", reduce = FALSE)
}
\arguments{
\item{data}{A dataframe with \code{datetime}, \code{stationname}, \code{value} and (salinity) \code{zone} columns. For example, the result
of \code{\link[=assignGeometry]{assignGeometry()}} or \code{\link[=assignZones]{assignZones()}}.}

\item{stat}{The statistics to aggregate data. Predifined option is "\code{mean}". See Details for the full list}

\item{reduce}{Logical. If \code{TRUE} merges individual statistics and simplifies. Default is \code{FALSE}.}
}
\value{
A dataframe.
}
\description{
Aggregate data for year range (max year to min year) per zone (level 4).
}
\details{
Available statistics are:
\itemize{
\item "\code{mean}" calculates Arithmetic Mean, uses \code{\link[base:mean]{base::mean()}},
\item "\code{sd}" calculates Stadard Deviation, uses \code{\link[stats:sd]{stats::sd()}}
\item "\code{se}" calculates Standard Error,
\item "\code{median}" calculates the Median Value. uses \code{\link[stats:median]{stats::median()}},
\item "\code{min}" calculates the minima, uses \code{\link[base:Extremes]{base::min()}},
\item "\code{max}" calculates the maxima, uses \code{\link[base:Extremes]{base::max()}},
\item "\verb{5perc}" calculates the 5th percentile,
\item "\verb{10perc}" calculates the 10th percentile,
\item "\verb{90perc}" calculates the 90th percentile,
\item "\verb{99perc}" calculates the 99th percentile,
\item "\code{sum}" calculates the sum, uses \code{\link[base:sum]{base::sum()}}
}
}
\examples{
# import data from high- and lowtide parameters
tide_parameters <- c(9694,9695)
# download data
abiotic_data <- importAbioticData(tide_parameters, start = 2012, end = 2015)

# assign geometry
abiotic_data_geo <- assignGeometry(abiotic_data, "T2021", TRUE)

# aggregate data for the whole period
abiotic_data_period_mean <- aggregatePeriodPerZone(abiotic_data_geo, reduce = TRUE)
}
