% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignRegion.R
\name{assignGeometry}
\alias{assignGeometry}
\title{Assign Scheldt geometries}
\usage{
assignGeometry(data, map, external)
}
\arguments{
\item{data}{A dataframe with 'latitude' and 'longitude' columns. For example, the result
of \code{\link[=importAbioticData]{importAbioticData()}} or \code{\link[=importBioticData]{importBioticData()}}}

\item{map}{Two options:
\itemize{
\item "T2015" to use the Scheldt Estuary shapefile of the T2015 report
\item "T2021" to use the Scheldt Estuary shapefile of the T2021 report
}}

\item{external}{Logical. Select \code{TRUE} to conditionally check stations outside standard geometry (like boundary stations)}
}
\value{
Input dataframe with extra columns added for \code{segment}, \code{zones} and \code{regions}.
}
\description{
\code{assignGeometry()} assigns Scheldt segments, zones and regions as extra columns of the input dataframe. The function uses
internal shapefiles (\code{map} argument) to intersect coordinate points from \code{data}
}
\details{
For more information about Scheldt segments, zones and regions, please visit the \href{https://www.scheldemonitor.be/sites/scheldemonitor.be/files/2023-08/Evaluatiemethodiek\%20Schelde-estuarium\%20Update\%202021.pdf}{T2021 SM report}
}
\examples{
# import data from high- and lowtide parameters
tide_parameters <- c(9694,9695)

# download data
abiotic_data <- importAbioticData(tide_parameters, start = 2010, end = 2015)

# add geometry
abiotic_data_geometry <- assignGeometry(abiotic_data, "T2015", external = TRUE)
}
