% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignRegion.R
\name{assignRegion}
\alias{assignRegion}
\title{Assign region (Westerschelde, Zeeschelde, ...)}
\usage{
assignRegion(data, map, external)
}
\arguments{
\item{data}{A dataframe with 'latitude' and 'longitude' columns. For example, the result
of \code{\link[=importAbioticData]{importAbioticData()}} or \code{\link[=importBioticData]{importBioticData()}}}

\item{map}{Two options:
\itemize{
\item "T2015" to use the Scheldt Estuary shapefile of the T2015 report
\item "T2021" to use the Scheldt Estuary shapefile of the T2021 report
}}

\item{external}{Logical. Select \code{TRUE} to conditionally check stations outside standard geometry (like boundary stations)}
}
\value{
Input dataframe with extra columns \code{segment} and \code{zone}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

Same functionality as \code{\link[=assignGeometry]{assignGeometry()}}, that we recommend to use instead
}
\examples{
# import data from high- and lowtide parameters
tide_parameters <- c(9694,9695) # High- and lowtide parameters in NAP+cm

# download data
abiotic_data <- importAbioticData(tide_parameters, start = 2010, end = 2012)

# assign zones
abiotic_data_region <- assignRegion(abiotic_data, "T2021", TRUE)
}
