% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignRegion.R
\name{assignZones}
\alias{assignZones}
\title{Assign salinity zones}
\usage{
assignZones(data, map, external)
}
\arguments{
\item{data}{A dataframe with 'latitude' and 'longitude' columns. For example, the result
of \code{\link[=importAbioticData]{importAbioticData()}} or \code{\link[=importBioticData]{importBioticData()}}}

\item{map}{Two options:
\itemize{
\item "T2015" to use the Scheldt Estuary shapefile of the T2015 report
\item "T2021" to use the Scheldt Estuary shapefile of the T2021 report
}}

\item{external}{Logical. Select \code{TRUE} to conditionally check stations outside standard geometry (like boundary stations)}
}
\value{
Input dataframe with extra columns \code{segment} and \code{zone}.
}
\description{
\code{assignZones()} assigns Scheldt segment and salinity zone as an extra column of the input dataframe. The function uses
the overarching function \code{\link[=assignGeometry]{assignGeometry()}}
}
\examples{
# import data from high- and lowtide parameters
tide_parameters <- c(9694,9695) # High- and lowtide parameters in NAP+cm

# download data
abiotic_data <- importAbioticData(tide_parameters, start = 2010, end = 2012)

# assign zones
abiotic_data <- assignZones(abiotic_data, "T2015", TRUE)
}
