% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateShannonIndex.R
\name{calculateShannonIndex}
\alias{calculateShannonIndex}
\title{Calculate Shannon index}
\usage{
calculateShannonIndex(value)
}
\arguments{
\item{value}{A column of a dataframe or a list of values.}
}
\value{
A double. The calculated Shannon Index for that set of values.
}
\description{
\code{calculateShannonIndex()} determines the Shannon index for a vector with values greater than 0. Value indicates abundance.
This can be e.g total number, density or coverage.
}
\examples{
# example with dataframe
# create a dataframe with species, and their densities:
df <- data.frame(species = c("s1", "s2", "s3"), value = 1:3)

# calculate index for the densities (value) in my dataset
shannonIndex <- scheldemonitoR::calculateShannonIndex(value = df$value)

# example with a list
density_values <- c(1,2,3,4)
shannonIndex <- scheldemonitoR::calculateShannonIndex(density_values)

# or directly
shannonIndex <- scheldemonitoR::calculateShannonIndex(c(1,2,3,4))
}
