% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignRegion.R
\name{checkBoundaries}
\alias{checkBoundaries}
\title{Check if a station is a boundary station}
\usage{
checkBoundaries(data, topic = "WK")
}
\arguments{
\item{data}{A dataframe with "omes", "segment", "zone" or "regions" columns. For example, the result
of \code{\link[=assignGeometry]{assignGeometry()}}, \code{\link[=assignOMES]{assignOMES()}}, \code{\link[=assignRegion]{assignRegion()}} or
\code{\link[=assignZones]{assignZones()}}}

\item{topic}{Three options for waterquiality: "WK", "WQ", "Waterkwaliteit"}
}
\value{
Dataframe
}
\description{
Checks if stations from a dataframe are considered a boundary station, and if so, reasign thems as boundary.
}
\examples{
# import data for 1046: 'Temperatuur in graden celcius in oppervlaktewater'
temperature_data <- importAbioticData("1046", start = 2018, end = 2023)

# assign zones
temperature_zone <- assignZones(temperature_data, "T2021", TRUE)

# check boundaries
temperature_boundaries <- checkBoundaries(temperature_zone, "WK")
}
