% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckDataAvailability.R
\name{checkDuplicateDate}
\alias{checkDuplicateDate}
\title{Check presence of duplicate datetimes for specific parameter or station(s)}
\usage{
checkDuplicateDate(
  data,
  station = NULL,
  visual = TRUE,
  pause = FALSE,
  dup_out = TRUE
)
}
\arguments{
\item{data}{A dataframe with \code{datetime}, \code{dataprovider}, and \code{stationname} columns. For example, the result
of \code{\link[=importAbioticData]{importAbioticData()}} or \code{\link[=importBioticData]{importBioticData()}}.}

\item{station}{Character or list. Station name or list of stationnames to check. Stations must be present in \code{data}.
If empty all distinct stationnames will be checked for duplicate dates. Default is \code{NULL}}

\item{visual}{Logical. If 'TRUE' a visual output will be generated. Default is \code{TRUE}.}

\item{pause}{Logical. If 'TRUE' function will ask whether to continue with the next station (when \code{station} > 1).
Default is \code{FALSE}.}

\item{dup_out}{Logical. If 'TRUE' returns a dataframe with potential duplicates. Default is \code{TRUE}.}
}
\value{
A dataframe and, if requested (\code{visual}) a plot
}
\description{
Check presence of duplicate datetimes for specific parameter or station(s)
}
\examples{
# 9694 is the parameter id for high tide in NAP
biotic_data_aphia <- importBioticData(129938, 1994, 2008, "aphia")

# rename datetime column
colnames(biotic_data_aphia)[colnames(biotic_data_aphia) == "observationdate"] <- "datetime"

# stations to check
stations <- c("WS Centr./West.deel WESTSDPLOT4_1212","WS Centr./West.deel WESTSDPLOT4_1213",
              "WS Centr./West.deel WESTSDPLOT4_1215", "WS Centr./West.deel WESTSDPLOT4_1221",
              "WS Centraal deel WESTSDMDN_0411", "WS Centraal deel WESTSDMDN_0433")

# get potential duplicates
duplicates <- checkDuplicateDate(biotic_data_aphia, station = stations,
                                 visual = TRUE, pause = TRUE, dup_out = TRUE)
}
