% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckDataAvailability.R
\name{depictDataAvailability}
\alias{depictDataAvailability}
\title{Define a graphical depiction of data availability (provider ans station)}
\usage{
depictDataAvailability(data, param = NULL, resolution = "year")
}
\arguments{
\item{data}{A dataframe with \code{datetime}, \code{dataprovider}, and \code{stationname} columns. For example, the result
of \code{\link[=importAbioticData]{importAbioticData()}} or \code{\link[=importBioticData]{importBioticData()}}.}

\item{param}{Character. See Details. Name of the parameter or list of parameters present in \code{data}, to add
to plot labels. Defaul is \code{NULL}.}

\item{resolution}{Either "year" or "month". "y" and "m" also accepted. Default is "year".}
}
\value{
A list of 2 plots. See Examples.
* First element plots available information per dataprovider.
* Second element plots available information per station.
}
\description{
Define a graphical depiction of data availability (provider ans station)
}
\details{
\code{depictDataAvailability()} does not check for unique parameternames. Recommendation is that \code{data} includes information for
just one parameter.
}
\examples{
# 9694 is the parameter id for high tide in NAP
tide_data <- importAbioticData(9694, start = 2018, end = 2021)

# plot available data
tide_availability_figures <- depictDataAvailability(tide_data, "high tide", "y")

# plot both plot elements
tide_availability_figures

# plot only available data per provider
tide_availability_figures[[1]]

# plot only available data per station
tide_availability_figures[[2]]
}
