% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateGraphics.R
\name{distToVlissingen}
\alias{distToVlissingen}
\title{Calculate distance to Vlissingen.}
\usage{
distToVlissingen(data)
}
\arguments{
\item{data}{A dataframe with 'latitude' and 'longitude' columns. For example, the result
of \code{\link[=importAbioticData]{importAbioticData()}} or \code{\link[=importBioticData]{importBioticData()}}}
}
\value{
A dataframe.
}
\description{
Uses a river network file creating using the \link{riverdist} pacakge. The funcion assigns the closes
distance to a segment and thn from that segment to Vlissingen for each of the coordinate points
from a dataframe.

Read more \href{https://jsta.github.io/riverdist/articles/riverdist_vignette.html}{here}.
}
\examples{
# import data for 1046: 'Temperatuur in graden celcius in oppervlaktewater'
temperature_data <- importAbioticData("1046", start = 2018, end = 2023)

# calculate distantce to Vlissingen:
temperature_data_vliss <- distToVlissingen(temperature_data)
}
