% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectData.R
\name{excludeExtremeValues}
\alias{excludeExtremeValues}
\title{Exclude extreme values}
\usage{
excludeExtremeValues(data, method = "iqr", limit = 3)
}
\arguments{
\item{data}{dataframe with \code{datetime} and \code{value} columns. For example, the result
of \code{\link[=importAbioticData]{importAbioticData()}} or \code{\link[=importBioticData]{importBioticData()}}.}

\item{method}{Available options "range", "min", "max", "pct" (perentile), "iqr" (Interquartile range).
See details. Default is \code{iqr}.}

\item{limit}{Numeric. Values higher than this number are considered extremes.}
}
\value{
A list of dataframes. The first one is the dataframe with extremes removed; the second is the
extreme values found.
}
\description{
Exclude extreme values.
}
\details{
\code{method = range} allows to input a value range for \code{limit}. For example \code{limit = c(1, 10)}.
}
\examples{
# import data from high- and lowtide parameters
tide_parameters <- c(9694,9695)

# download data
abiotic_data <- importAbioticData(tide_parameters, start = 2018, end = 2019)

# exclude values lower than 0:
data_extreme <- excludeExtremeValues(abiotic_data, method = "min", limit = 0)

# exclude values by range: lower than 1 and higher than 400
data_extreme <- excludeExtremeValues(abiotic_data,
                                     method = "range", limit = c(1, 400))
}
