% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectData.R
\name{exploreDataStation}
\alias{exploreDataStation}
\title{Make station-specific temporal plot}
\usage{
exploreDataStation(
  data,
  varname = "Variable",
  station = NULL,
  plottype = "line"
)
}
\arguments{
\item{data}{A dataframe with 'datetime', 'value' and 'stationname' columns. For example, the result
of \code{\link[=importAbioticData]{importAbioticData()}} or \code{\link[=importBioticData]{importBioticData()}}.}

\item{varname}{Character. Optional. Variable name to include in the title of the plot. Placehorder is \code{Variable}.}

\item{station}{Character or list. Station or list of stations to plot. If \code{NULL} the first station that appears
in the dataframe will be selected.  is \code{NULL}.}

\item{plottype}{Character. Options are \code{line} or \code{boxplot}. Dafualt is \code{line}.}
}
\value{
A ggplot object.
}
\description{
Make station-specific temporal plot.
}
\examples{
# make a list of the dataset ids
imis_datasets <- c(1073, 5191)

# download data
biotic_data_imis <- importBioticData(imis_datasets, 2015, 2021, 'imis')

# correct datetime colum name
colnames(biotic_data_imis)[colnames(biotic_data_imis) == "observationdate"] <- "datetime"

# make temporal plot
plot <- exploreDataStation(biotic_data_imis, "Densiteit van zooplankton (ind/m3)", "line")
plot
}
