% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleGraphicsHelpers.R
\name{exportPlot}
\alias{exportPlot}
\title{Export plot as an image (png)}
\usage{
exportPlot(p, fname, w = 16, h = 10)
}
\arguments{
\item{p}{Plot object. For example the result of \code{\link[=exploreDataStation]{exploreDataStation()}}.}

\item{fname}{Character. Filename.}

\item{w}{Numeric. Desired width of the plot. Default = \code{16}.}

\item{h}{Numeric. Desired height of the plot. Default = \code{10}}
}
\value{
A plot saved to current directory.
}
\description{
Export plot as an image (png)
}
\examples{
# make a list of the dataset ids
imis_datasets <- c(1073, 5191)

# download data
biotic_data_imis <- importBioticData(imis_datasets, 2015, 2021, 'imis')

# correct datetime colum name
colnames(biotic_data_imis)[colnames(biotic_data_imis) == "observationdate"] <- "datetime"

# make temporal plot
plot <- exploreDataStation(biotic_data_imis, "Densiteit van zooplankton (ind/m3)", "line")
exportPlot(plot, "densityplot.png") # save to current directory

# save to specific directory
# mydirectory <- "/Mydirectory"
# exportPlot(file.path(mydirectory, "densityplot.png"))
}
