% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleGraphicsHelpers.R
\name{firstUp}
\alias{firstUp}
\title{Change first charachter of a string to uppercase}
\usage{
firstUp(x)
}
\arguments{
\item{x}{character.}
}
\value{
character.
}
\description{
Change first charachter of a string to uppercase
}
\examples{
firstUp("scheldemonitor")
}
