% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckDataAvailability.R
\name{heatmapDataAvailability}
\alias{heatmapDataAvailability}
\title{Describe data availability through heatmap}
\usage{
heatmapDataAvailability(
  data,
  param = "parameter",
  resolution = "year",
  type = "station"
)
}
\arguments{
\item{data}{A dataframe with \code{datetime}, \code{dataprovider}, and \code{stationname} columns. For example, the result
of \code{\link[=importAbioticData]{importAbioticData()}} or \code{\link[=importBioticData]{importBioticData()}}.}

\item{param}{Character. See Details. Name of the parameter or list of parameters present in \code{data}, to add
to plot labels. Defaul is \code{NULL}.}

\item{resolution}{Either "year" or "month". "y" and "m" also accepted. Default is "year".}

\item{type}{Either "station" or "dataprovider". Default is "station".}
}
\value{
A heatmap.
}
\description{
Describe data availability through heatmap, with option to choose to use dataprovider or stations.
}
\details{
\code{heatmapDataAvailability()} does not check for unique parameternames. Recommendation is that \code{data} includes information for
just one parameter.
}
\examples{
# 9694 is the parameter id for high tide in NAP
tide_data <- importAbioticData(9694, start = 2018, end = 2021)

# plot available data
tide_availability_figures <- heatmapDataAvailability(tide_data, "high tide", "year", "station")

# Example with abiotic data ---------------------------------------------------------
# abiotic data #949 & 1074
fytoplankton_data <- importBioticData(1074, start = 2013, end = 2015, source = "imis")

# correct datetime columname
colnames(fytoplankton_data)[colnames(fytoplankton_data) == "observationdate"] <- "datetime"

# plot available data
heatmap_fyto <- heatmapDataAvailability(fytoplankton_data, "high tide", "year", "station")
}
