% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.R
\name{importAbioticData}
\alias{importAbioticData}
\title{Import Abiotic Data}
\usage{
importAbioticData(parameter, start, end)
}
\arguments{
\item{parameter}{A ScheldeMonitor parameter id or a list of parameter ids}

\item{start}{Start year}

\item{end}{End year}
}
\value{
A dataframe with the available data for the parameters and years selected
}
\description{
\code{importAbioticData()} uses the \href{https://www.scheldemonitor.be/dataproducts/en/download/}{ScheldeMonitor data download webservices}
to download abiotic data, returning a dataframe. To download biotic data please use \code{\link[=importBioticData]{importBioticData()}}
}
\examples{
# Define parameters
tide_parameters <- c(9694,9695) # High- and lowtide parameters in NAP+cm

# download data
abiotic_data <- importAbioticData(tide_parameters, 2010, 2023)
}
