% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.R
\name{importBioticData}
\alias{importBioticData}
\title{Import Biotic Data}
\usage{
importBioticData(parameter, start, end, source)
}
\arguments{
\item{parameter}{Two options:
\itemize{
\item The id or list of ids of an \href{https://www.vliz.be/nl/imis?module=dataset}{IMIS (Integrated Marine Information System) dataset}.
\item The AphiaID or list of AphiaIDs of a scientific name. This information is available in the \href{(https://www.marinespecies.org/index.php)}{WoRMS website} or through the \code{\link[worrms:wm_name2id]{worrms::wm_name2id()}} function.
}}

\item{start}{Start year}

\item{end}{End year}

\item{source}{Two options:
\itemize{
\item \code{"imis"} if \code{parameter} is an IMIS dataset id.
\item \code{"aphia"} if \code{parameter} is an AphiaID
}}
}
\value{
A dataframe with the available data for the parameters and years selected
}
\description{
\code{importBioticData()} uses the \href{https://www.scheldemonitor.be/dataproducts/en/download/}{ScheldeMonitor data download webservices}
to download biotic data, returning a dataframe. To download abiotic data please use \code{\link[=importAbioticData]{importAbioticData()}}
}
\examples{
# using source = 'imis' ---------------------------------------------------------------------
# Define parameters (IMIS dataset ids):
# 1073: OMES: Monitoring zooplankton in the Zeeschelde
## (https://www.vliz.be/nl/imis?module=dataset&dasid=1073)
# 5191: MWTL biological monitoring network Westerschelde: Mesozooplankton
## (https://www.vliz.be/nl/imis?module=dataset&dasid=5191)

# make a list of the dataset ids
imis_datasets <- c(1073, 5191)

# download data
biotic_data_imis <- importBioticData(imis_datasets, 2015, 2021, 'imis')

# using source = 'aphia' --------------------------------------------------------------------
# find AphiaID for bristleworm
# install.packages(worrms)

# find the AphiaID from the taxonomic name:
worrms::wm_name2id("Aphelochaeta marioni")

# dowloand data using the AphiaID
biotic_data_aphia <- importBioticData(129938, 1994, 2008, "aphia")
}
