% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportMDA.R
\name{importMDAFile}
\alias{importMDAFile}
\title{Import file from MDA}
\usage{
importMDAFile(username, password, file_id, sep = ";")
}
\arguments{
\item{username}{Character. Username (MDA account) to use for the connection.}

\item{password}{Character. Password (MDA account) to use for the connection.}

\item{file_id}{Character. MDA id of the file to download. See 'Details'.}

\item{sep}{Character. If known. Single character use to separate fields within a record.
Options are: \code{","}, \code{";"}, \code{"."}, \code{" "}, \code{""}, \code{"\\t"}.}
}
\value{
A data object.
}
\description{
Imports a file from a file stored in \href{https://mda.vliz.be/}{Marine Data Archive} (MDA).
Previous registration to the MDA is needed, as the funtion uses MDA credential to access data that
is either publically available or that belongs to the MDA user.
}
\details{
You will find the MDA file ID by checking the file 'Metadata'. in 'Direc link', evereything starting from
\code{VLIZ_} is the file id.
For example for \href{https://mda.vliz.be/print.php?file=VLIZ_00000308_6422b9034a3fe367642157}{this file},
the file id is: \code{VLIZ_00000308_6422b9034a3fe367642157}.
}
\examples{
\dontshow{
username <- Sys.getenv("mda_userid_gmail")
password <- Sys.getenv("mda_password_gmail")
}
\dontrun{
username <- "myemail@email.com"
password <- "mysafepasswd"
}

# file id (See 'Details'.)
MDA_file_id <- "VLIZ_00000308_6422b9034a3fe367642157"

# get file
traits <- importMDAFile(username = username, password = password, file_id = MDA_file_id)
}
