% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateGraphics.R
\name{makePeriodPlot}
\alias{makePeriodPlot}
\title{Make barplot for reporting period}
\usage{
makePeriodPlot(
  data,
  locations = NULL,
  varname = "Variable",
  varunit = "-",
  threshold = NULL
)
}
\arguments{
\item{data}{A dataframe with columns \code{location} (as returned with \code{\link[=aggregatePeriodPerStation]{aggregatePeriodPerStation()}}),
\code{period} (year period) and \code{value} (parameter value).}

\item{locations}{Character or list. List of locations to plot. If null,}

\item{varname}{Character. Optional. Variable name to include in the title of the plot. Placehorder is \code{Variable}.}

\item{varunit}{Character. Optional. Variable unit to include in the title of the plot. Placehorder is \code{-}.}

\item{threshold}{Numeric. Add a threshold line. Default is \code{NULL}.}
}
\value{
A plot.
}
\description{
Make barplot for reporting period

Check also: `04_Waterkwaliteit/03_Toetsparameter Nutrienten/Scripts/c. Analysis scripts/AnalyseNitrietPeriodiek.R``
}
\examples{
# import data for 1046: 'Temperatuur in graden celcius in oppervlaktewater'
temperature_data <- importAbioticData("1046", start = 2018, end = 2023)

# add geometries
temperature_data_geo <- assignGeometry(temperature_data, "T2015", external = TRUE)

# aggregate data per station
df <- aggregatePeriodPerStation(temperature_data_geo, stat = 'max', reduce = TRUE)

# add period. Data can also be filtered for a specific period (2018-2020)
# using scheldemonitoR aggregate functions
df$period <- paste0("2018-2023")

makePeriodPlot(df, location = c("154100","157000","159000", "160200"), "temperature", "C", 18)
}
