% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateGraphics.R
\name{makeSurfacePlot}
\alias{makeSurfacePlot}
\title{Make surface plot (x = date, y = distance Vlissingen, z = value parameter)}
\usage{
makeSurfacePlot(
  data,
  varname = "Variable",
  varunit = "-",
  maxgapdays = 80,
  colclass = 10
)
}
\arguments{
\item{data}{A dataframe at least with 'latitude', 'longitude' and 'datetime' columns. For example, the result
of \code{\link[=importAbioticData]{importAbioticData()}} or \code{\link[=importBioticData]{importBioticData()}}. Preferred, a dataframe with
column 'distance' as the result of \code{\link[=distToVlissingen]{distToVlissingen()}}}

\item{varname}{Character. Optional. Variable name to include in the title of the plot. Placehorder is \code{Variable}.}

\item{varunit}{Character. Optional. Variable unit to include in the title of the plot. Placehorder is \code{-}.}

\item{maxgapdays}{Numeric. Maximum desired gap days.}

\item{colclass}{Numeric. Break points for layer.}
}
\value{
A plot object.
}
\description{
Make surface plot (x = date, y = distance Vlissingen, z = value parameter)
}
\examples{
# import data for 1046: 'Temperatuur in graden celcius in oppervlaktewater'
temperature_data <- importAbioticData("1046", start = 2018, end = 2023)

# calculate distantce to Vlissingen:
temperature_data_vliss <- distToVlissingen(temperature_data)

# With column distance calculated ------------------------------------------
surface_plot <- makeSurfacePlot(temperature_data_vliss, varname = "Temperatuur", varunit = "C")
# Without column distance calculated ---------------------------------------
surface_plot <- makeSurfacePlot(temperature_data, varname = "Temperatuur", varunit = "C")
}
