% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveOutput.R
\name{savePeriodWide}
\alias{savePeriodWide}
\title{Saves aggregated by Period data}
\usage{
savePeriodWide(data, locations = NULL, cols = NULL, fileout)
}
\arguments{
\item{data}{A dataframe with columns 'location', 'period' and 'value' columns. Like
the result of \code{\link[=aggregatePeriodPerZone]{aggregatePeriodPerZone()}}.}

\item{locations}{Character or list. Spatial resolution to filter columbn location.
Default is \code{NULL}.}

\item{cols}{Character or list. Columns to select from dataframe.}

\item{fileout}{Path ans extension to save the file to. For example \code{Plots/myfile.txt}.}
}
\value{
A file.
}
\description{
Saves aggregated by Period data.
}
\examples{
library(dplyr)
# import data
temperature_data <- importAbioticData("1046", start = 2018, end = 2023)
# assign geometry
temperature_data_geo <- assignGeometry(temperature_data, "T2015", external = TRUE)
# aggregate data per period
temperature_data_geo_period <- aggregatePeriodPerZone(temperature_data_geo, stat = "mean")
# rename columns
temperature_data_geo_period <- temperature_data_geo_period \%>\%
                                        rename(location = zone, period = data)
# save info
savePeriodWide(temperature_data_geo_period, fileout = "mifile.txt")
}
