% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectData.R
\name{selectTimePeriod}
\alias{selectTimePeriod}
\title{Select data for a time period}
\usage{
selectTimePeriod(data, start, end)
}
\arguments{
\item{data}{A dataframe with a \code{datetime} column. For example, the result of \code{\link[=importAbioticData]{importAbioticData()}}
or \code{\link[=importBioticData]{importBioticData()}}.}

\item{start}{Numeric. Start year.}

\item{end}{Numeric. End year.}
}
\value{
A dataframe.
}
\description{
Select data based on input start year and end year.
}
\examples{
# import data from high- and lowtide parameters
tide_parameters <- c(9694,9695)

# download data
abiotic_data <- importAbioticData(tide_parameters, start = 2018, end = 2021)

# Get the data for a specific time period:
abiotic_data_selected_years <- selectTimePeriod(abiotic_data, 2019, 2019)
}
