% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckDataAvailability.R
\name{summarizeAvailableDatabyYear}
\alias{summarizeAvailableDatabyYear}
\title{Summarize Available data by year}
\usage{
summarizeAvailableDatabyYear(data, start, end)
}
\arguments{
\item{data}{A dataframe with \code{datetime}, \code{parametername}, and \code{standardparameterid} columns. For example, the result
of \code{\link[=importAbioticData]{importAbioticData()}} or \code{\link[=importBioticData]{importBioticData()}}.}

\item{start}{Begin year.}

\item{end}{End year.}
}
\value{
A dataframe.
}
\description{
Summarize Available data by year
}
\examples{
# import data from high- and lowtide parameters
tide_parameters <- c(9694,9695)

# download data
abiotic_data <- importAbioticData(tide_parameters, start = 2018, end = 2021)

available_data <- summarizeAvailableDatabyYear(abiotic_data, 2020, 2021)
}
