% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleGraphicsHelpers.R
\name{themeSib}
\alias{themeSib}
\title{Apply theme of 'Schelde in beeld' (sib)}
\usage{
themeSib()
}
\value{
plot object.
}
\description{
Apply theme of 'Schelde in beeld' (sib)
}
\examples{
# make a list of the dataset ids
imis_datasets <- c(1073, 5191)

# download data
biotic_data_imis <- importBioticData(imis_datasets, 2015, 2021, 'imis')

# correct datetime colum name
colnames(biotic_data_imis)[colnames(biotic_data_imis) == "observationdate"] <- "datetime"

# make temporal plot
plot <- exploreDataStation(biotic_data_imis, "Densiteit van zooplankton (ind/m3)", "line")
plotsib <- plot + themeSib()
plotsib
}
