% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectData.R
\name{updateStationname}
\alias{updateStationname}
\title{Update stationnames}
\usage{
updateStationname(data)
}
\arguments{
\item{data}{A dataframe with 'stationname' column. For example, the result
of \code{\link[=importAbioticData]{importAbioticData()}} or \code{\link[=importBioticData]{importBioticData()}}}
}
\value{
A dataframe.
}
\description{
Update old stationnames to new stationnames. For example \code{211900} to \verb{Dijle Zenne (211900)}.
}
\examples{
# import data from high- and lowtide parameters
tide_parameters <- c(9694,9695)

# download data
abiotic_data <- importAbioticData(tide_parameters, start = 2018, end = 2020)

# rename stations
abiotic_data_updated <- updateStationname(abiotic_data)
}
